==> 4 - Dummies Departamentos.ipynb <==
# Desafío N° 2 
## 4 - Construccion de Dummies para Departamentos

### Grupo N° 5

    Integrantes:
                Arangue, Marcelo            
                Bardauil, Joaquín              
                Marquez, Hector              
                Neustadt, Alejandro       
                Pero, Felipe                                    


# Importamos las librerías que usaremos durante la limpieza:
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
import re
import scipy.stats as sts
from bokeh.plotting import figure
from bokeh.io import output_notebook, show
from IPython.display import display

df_properatti_depto = pd.read_csv("prop_apartament_1.csv")

from sklearn.feature_extraction.text import CountVectorizer

#DUMMY BALCON

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="balcon", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_balcon = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_balcon, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="balcon", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_balcon2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_balcon2, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera y la segunda) en una sola columna nueva
df_properatti_depto["balcon"] = df_properatti_depto["balcon_x"]|df_properatti_depto["balcon_y"]
##Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["balcon_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["balcon_y"], axis = 1, inplace = True) 

#DUMMY CENTRICO

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(centrico|centro)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_centrico = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_centrico, on = df_properatti_depto.index)
df_properatti_depto["centrico_x"] = df_properatti_depto["centrico"]|df_properatti_depto["centro"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["centrico"], axis = 1, inplace = True) 
df_properatti_depto.drop(["centro"], axis = 1, inplace = True) 

#Creo la segunda columna para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(centrico|centro)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_centrico2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_centrico2, on = df_properatti_depto.index)
df_properatti_depto["centrico_y"] = df_properatti_depto["centrico"]|df_properatti_depto["centro"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["centrico"], axis = 1, inplace = True) 
df_properatti_depto.drop(["centro"], axis = 1, inplace = True)

#Junto las dos(la primera y la segunda) en una sola columna nueva, en la que habra un 1 si se encontró coincidencia en al menos una de las columnas de las anteriores) 
df_properatti_depto["centrico"] = df_properatti_depto["centrico_x"]|df_properatti_depto["centrico_y"]
#Dropeo las dos que use, dejando solo la nueva(en la cual busque las palabras frente y vista al mar en description y en title)
df_properatti_depto.drop(["centrico_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["centrico_y"], axis = 1, inplace = True) 


#DUMMY AIRE ACONDICIONADO

#Creo la primera columna para description
#vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="aire acondicionado", analyzer="word", binary=True)
#X = vec.fit_transform(df_properatti_depto["description"])
#dummy_aire_acondicionado = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
#df_properatti_depto = pd.merge(df_properatti_depto, dummy_aire_acondicionado, on = df_properatti_depto.index)
#Dropeo
#df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
#vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="aire acondicionado", analyzer="word", binary=True)
#X = vec.fit_transform(df_properatti_depto["title"])
#dummy_aire_acondicionado2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
#df_properatti_depto = pd.merge(df_properatti_depto, dummy_aire_acondicionado2, on = df_properatti_depto.index)
#Dropeo
#df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
#df_properatti_depto["aire acondicionado"] = df_properatti_depto["aire acondicionado_x"]|df_properatti_depto["aire acondicionado_y"]
#Dropeo las dos que use, dejando solo la nueva
#df_properatti_depto.drop(["aire acondicionado_x"], axis = 1, inplace = True) 
#df_properatti_depto.drop(["aire acondicionado_y"], axis = 1, inplace = True) 

#DUMMY BAULERA

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="baulera", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_baulera = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_baulera, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="baulera", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_baulera2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_baulera2, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
df_properatti_depto["baulera"] = df_properatti_depto["baulera_x"]|df_properatti_depto["baulera_y"]
#Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["baulera_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["baulera_y"], axis = 1, inplace = True) 

#DUMMY TORRE

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="torre", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_torre = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_torre, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="torre", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_torre2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_torre2, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
df_properatti_depto["torre"] = df_properatti_depto["torre_x"]|df_properatti_depto["torre_y"]
#Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["torre_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["torre_y"], axis = 1, inplace = True) 

#DUMMY COCHERA

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(cochera|garage|guardacoche)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_cochera = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_cochera, on = df_properatti_depto.index)
df_properatti_depto["Cochera"] = df_properatti_depto["cochera"]|df_properatti_depto["garage"]|df_properatti_depto["guardacoche"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["cochera"], axis = 1, inplace = True) 
df_properatti_depto.drop(["garage"], axis = 1, inplace = True) 
df_properatti_depto.drop(["guardacoche"], axis = 1, inplace = True)



#DUMMY PILETA

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(pileta|piscina)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_pileta = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_pileta, on = df_properatti_depto.index)
df_properatti_depto["pileta_x"] = df_properatti_depto["pileta"]|df_properatti_depto["piscina"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["pileta"], axis = 1, inplace = True) 
df_properatti_depto.drop(["piscina"], axis = 1, inplace = True) 

#Creo la segunda columna para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(pileta|piscina)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_pileta2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_pileta2, on = df_properatti_depto.index)
df_properatti_depto["pileta_y"] = df_properatti_depto["pileta"]|df_properatti_depto["piscina"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["pileta"], axis = 1, inplace = True) 
df_properatti_depto.drop(["piscina"], axis = 1, inplace = True)

#Junto las dos(la primera y la segunda) en una sola columna nueva, en la que habra un 1 si se encontró coincidencia en al menos una de las columnas de las anteriores) 
df_properatti_depto["pileta"] = df_properatti_depto["pileta_x"]|df_properatti_depto["pileta_y"]
#Dropeo las dos que use, dejando solo la nueva(en la cual busque las palabras frente y vista al mar en description y en title)
df_properatti_depto.drop(["pileta_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["pileta_y"], axis = 1, inplace = True)

#DUMMY GIMNASIO 

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(gimnasio|gym)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_gimnasio = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_gimnasio, on = df_properatti_depto.index)
df_properatti_depto["Gimnasio"] = df_properatti_depto["gimnasio"]|df_properatti_depto["gym"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["gimnasio"], axis = 1, inplace = True) 
df_properatti_depto.drop(["gym"], axis = 1, inplace = True) 


#DUMMY SEGURIDAD

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="seguridad", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_seguridad = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_seguridad, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#DUMMY PARRILLA

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="parrilla", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_parrilla = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_parrilla, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="parrilla", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_parrilla2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_parrilla2, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
df_properatti_depto["parrilla"] = df_properatti_depto["parrilla_x"]|df_properatti_depto["parrilla_y"]
#Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["parrilla_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["parrilla_y"], axis = 1, inplace = True) 

#DUMMY SUM
#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern=" sum ", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_sum = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_sum, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 



#DUMMY SOLARIUM

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="solarium", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_solarium = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_solarium, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 


#DUMMY JARDIN
#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(jardin|parque|patio)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_jardin = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_jardin, on = df_properatti_depto.index)
df_properatti_depto["jardin_x"] = df_properatti_depto["jardin"]|df_properatti_depto["parque"]|df_properatti_depto["patio"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["jardin"], axis = 1, inplace = True) 
df_properatti_depto.drop(["parque"], axis = 1, inplace = True) 
df_properatti_depto.drop(["patio"], axis = 1, inplace = True)

#Creo la segunda columna para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="(jardin|parque|patio)", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_jardin2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_jardin2, on = df_properatti_depto.index)
df_properatti_depto["jardin_y"] = df_properatti_depto["jardin"]|df_properatti_depto["parque"]|df_properatti_depto["patio"]
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)
df_properatti_depto.drop(["jardin"], axis = 1, inplace = True) 
df_properatti_depto.drop(["parque"], axis = 1, inplace = True) 
df_properatti_depto.drop(["patio"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
df_properatti_depto["jardin"] = df_properatti_depto["jardin_x"]|df_properatti_depto["jardin_y"]
#Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["jardin_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["jardin_y"], axis = 1, inplace = True) 

#DUMMY AMENITIES

#Creo la primera columna para description
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="amenities", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["description"])
dummy_amenities = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_amenities, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True) 

#Creo la segunda para title
vec = CountVectorizer(strip_accents="unicode", lowercase=True, token_pattern="amenities", analyzer="word", binary=True)
X = vec.fit_transform(df_properatti_depto["title"])
dummy_amenities2 = pd.DataFrame(data = X.todense(), columns = vec.get_feature_names())
df_properatti_depto = pd.merge(df_properatti_depto, dummy_amenities2, on = df_properatti_depto.index)
#Dropeo
df_properatti_depto.drop(["key_0"], axis = 1, inplace = True)

#Junto las dos(la primera x y la segunda y) en una sola columna nueva
df_properatti_depto["amenities"] = df_properatti_depto["amenities_x"]|df_properatti_depto["amenities_y"]
#Dropeo las dos que use, dejando solo la nueva
df_properatti_depto.drop(["amenities_x"], axis = 1, inplace = True) 
df_properatti_depto.drop(["amenities_y"], axis = 1, inplace = True)

#Junto la columna amenities con las amenities anteriores, obteniendo la columna "Amenities"(con mayuscula)
df_properatti_depto["Amenities"] = df_properatti_depto["amenities"]|df_properatti_depto["pileta"]|df_properatti_depto["Gimnasio"]|df_properatti_depto["seguridad"]|df_properatti_depto["solarium"]|df_properatti_depto[" sum "]|df_properatti_depto["jardin"]|df_properatti_depto["parrilla"]
#Dropeo
df_properatti_depto.drop(["amenities"], axis = 1, inplace = True)

df_properatti_depto.columns

df_properatti_depto.shape

df_properatti_depto.head()

# Genero un nuevo dataframe para ser usado estandarizado en otras regreciones o con campos completos, con las dummies
df_properatti_depto.to_csv('prop_apartament_dummies.csv')
