==> 5 - Regresión lineal, Ridge y LASSO Departamentos.ipynb <==
# Desafío N° 2 
## 5 - Machine Learning

### Grupo N° 5

    Integrantes:
                Arangue, Marcelo            
                Bardauil, Joaquín              
                Marquez, Hector              
                Neustadt, Alejandro       
                Pero, Felipe                                    

%matplotlib inline
import random
import numpy as np
from matplotlib import pyplot as plt
from sklearn import datasets, linear_model

import pandas as pd
from scipy import stats
import seaborn as sns
from sklearn import datasets
from sklearn.model_selection import train_test_split, KFold, cross_val_score
from sklearn.linear_model import LinearRegression, Lasso, LassoCV, Ridge, RidgeCV
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.metrics import mean_squared_error, r2_score

df_dpto_connulos = pd.read_csv("prop_apartament_dummies.csv", usecols=[7,8,10,23,24,26,27,35])
df_dpto_connulos.columns

df_dpto = df_dpto_connulos.dropna()
df_dpto.shape

df_dpto['surface_covered_in_m2 * rooms'] = df_dpto['surface_covered_in_m2'] * df_dpto['rooms']

df_dpto.columns

X = df_dpto[['surface_covered_in_m2', 'rooms','surface_covered_in_m2 * rooms','balcon','centrico','torre','Cochera','Amenities']]
y = df_dpto.price_aprox_usd

Tengo que elegir entre los tres modelos(regresión lineal, regresion regularizada Ridgea y regresio regularizada LASSO)

# Primero, dividimos el dataser en training/validacion y test
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=123)
print(X_train.shape, y_train.shape)
print(X_test.shape, y_test.shape)

# habia pensado en separar el Xtest e ytest, pero como la funcion esta inicializada siempre da el mismo random, entonces
# no mete el test en el training generando overfiting
X_test.head()

# Ahora bien, vamos a realizar el proceso de tunning de los hiperparámetros de los modelos en base a validación cruzada.
# Por eso, instanciamos previamente tres objetos:
al_ridge = np.linspace(0.001, 0.5, 500)  
al_lasso = np.linspace(0.001, 0.5, 500)
kf = KFold(n_splits=5, shuffle=True)

# Instanciamos los modelos:
lm = LinearRegression()
lm_ridge_cv= RidgeCV(alphas=al_ridge, cv=kf, normalize=True)
lm_lasso_cv = LassoCV(alphas=al_lasso, cv=kf, normalize=True)
elastic_net = linear_model.ElasticNet(alpha=0.5, normalize=True)

# Hacemos los fits respectivos
lm.fit(X_train, y_train)

lm_ridge_cv.fit(X_train, y_train)

lm_lasso_cv.fit(X_train, y_train)

elastic_net.fit(X_train, y_train)

# Calculamos el R2 y el RMSE para ver la performance tiene cada modelo en el Training Set 

print(" Score Train Lineal: %.2f\n" % lm.score(X_train, y_train),
      "Score Train Ridge : %.2f\n" % lm_ridge_cv.score(X_train, y_train),
      "Score Train Lasso : %.2f\n" %  lm_lasso_cv.score(X_train, y_train),
      "Score Train Elastic : %.2f\n" % elastic_net.score(X_train, y_train))

y_pred_tr_lm = lm.predict(X_train)
y_pred_tr_ridge = lm_ridge_cv.predict(X_train)
y_pred_tr_lasso = lm_lasso_cv.predict(X_train)
y_pred_tr_elast = elastic_net.predict(X_train)

rmse = lambda y, y_pred: np.sqrt(mean_squared_error(y, y_pred))

print(" Train RMSE lineal   : %.2f \n" % rmse(y_train,y_pred_tr_lm),
      "Train RMSE Ridge    : %.2f \n" % rmse(y_train,y_pred_tr_ridge),
      "Train RMSE Lasso    : %.2f \n" % rmse(y_train,y_pred_tr_lasso),
     "Train RMSE Elastic    : %.2f \n" % rmse(y_train,y_pred_tr_elast))

# Lo único que queda es evaluar la performance del modelo. 
# Para ello debemos usar datos que no hayan formado parte del proceso de entrenamiento anterior.
# Entonces usamos el Test Set. 
# Hacemos las predicciones sobre la matriz de predictores del Test Set

y_pred_lm = lm.predict(X_test)
y_pred_ridge = lm_ridge_cv.predict(X_test)
y_pred_lasso = lm_lasso_cv.predict(X_test)
y_pred_elast = elastic_net.predict(X_test)

print(" Score Test Lineal: %.2f\n" % lm.score(X_test, y_test),
      "Score Test Ridge : %.2f\n" % lm_ridge_cv.score(X_test, y_test),
      "Score Test Lasso : %.2f\n" %  lm_lasso_cv.score(X_test, y_test),
      "Score Test Elastic : %.2f\n" % elastic_net.score(X_test, y_test))

# Testeo final del modelo sobre Test Set

print(" Test RMSE lineal= %.2f\n" % rmse(y_test, y_pred_lm),
      "Test RMSE Ridge = %.2f\n" %  rmse(y_test, y_pred_ridge),
      "Test RMSE Lasso = %.2f\n" %  rmse(y_test, y_pred_lasso),
     "Test RMSE Elastic = %.2f\n" % rmse(y_test,y_pred_elast))

Para saber con cual me quedo, comparo los Score Train y Test, con sus respectivos RMSE

## Regularización: Lasso

model = linear_model.LassoCV(cv=3).fit(X_train, y_train)
scores = cross_val_score(model, X_train, y_train, cv=kf, scoring='r2')
dict(alpha=model.alpha_, scores=scores, mean_score=scores.mean(), zero_coefs=(model.coef_ == 0).sum())

model = linear_model.LassoCV(cv=3, normalize=True).fit(X, y)
scores = cross_val_score(model, X, y, cv=kf, scoring='r2')
dict(alpha=model.alpha_, scores=scores, mean_score=scores.mean(), zero_coefs=(model.coef_ == 0).sum())

model = linear_model.RidgeCV(normalize=True).fit(X, y)
scores = cross_val_score(model, X, y, cv=kf, scoring='r2')
dict(alpha=model.alpha_, scores=scores, mean_score=scores.mean())
