==> 7 - Regresiones condicionales Departamentos.ipynb <==
# Desafío N° 2 
## 7 - Machine Learning

### Grupo N° 5

    Integrantes:
                Arangue, Marcelo            
                Bardauil, Joaquín              
                Marquez, Hector              
                Neustadt, Alejandro       
                Pero, Felipe   

import numpy as np
import pandas as pd
from scipy import stats
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn import datasets, linear_model
from sklearn.model_selection import train_test_split, KFold, cross_val_score
from sklearn.linear_model import LinearRegression, Lasso, LassoCV, Ridge, RidgeCV
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.metrics import mean_squared_error, r2_score

Usaré prop_apartament_dummies.csv, le haré el drop de nulos y crearé la nueva feature que resultan en prop_apartament_regresion. Lo hago así para obtener también las columnas state_name, ciudad y barrio, que servirán para correr las regresiones relativo a dichas zonas

df = pd.read_csv("prop_apartament_dummies.csv")
columnas=['surface_covered_in_m2', 'rooms', 'price_aprox_usd', 'balcon', 'centrico', 'torre', 'Cochera', 'Amenities']
df_drop = df[columnas].dropna()
indices = df_drop.index
df_dpto = df.iloc[indices]
df_dpto['surface_covered_in_m2 * rooms'] = df_dpto['surface_covered_in_m2'] * df_dpto['rooms']

print(df.shape)
print(df_dpto.shape)

df_dpto[columnas].head()

# Regresiones

El objetivo es automatizar el proceso de obtener modelos dados conjuntos de train y test. El valor final que queremos obtener es un DataFrame en donde cada fila represente a un modelo, junto con su score y su error cuadrático medio

En primer lugar, separamos los conjuntos de train y test y observamos la forma de cada uno

predictoras = ['surface_covered_in_m2', 'rooms', 'surface_covered_in_m2 * rooms', 'balcon', 'centrico', 'torre', 'Cochera', 'Amenities']

X = df_dpto[predictoras]
y = df_dpto.price_aprox_usd

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=123)

conjuntos = [X_train, X_test, y_train, y_test]
indices = [x.index for x in conjuntos]

def formas(X_train, X_test, y_train, y_test):
    print('Test')
    print(X_train.shape, np.array(y_train).ravel().shape)
    print('\n')
    print('Train')
    print(X_test.shape, y_test.shape)
    
formas(X_train, X_test, y_train, y_test)

X_test.head()

Definimos la función ```regresiones```, que genera modelos de regresión lineal, regresión Ridge y regresión Lasso a partir de dos conjuntos X_train e y_train

def regresiones(X_train, y_train):    
    # armamos los parametros para las regresiones
    al_ridge = np.linspace(0.001, 0.5, 500)  
    al_lasso = np.linspace(0.001, 0.5, 500)
    kf = KFold(n_splits=5, shuffle=True)
    
    # instanciamos los objetos correspondientes a cada regresion
    lm = LinearRegression()
    lm_ridge_cv= RidgeCV(alphas=al_ridge, cv=kf, normalize=False)
    lm_lasso_cv = LassoCV(alphas=al_lasso, cv=kf, normalize=False)
    elastic_net = linear_model.ElasticNet(alpha=0.5, normalize=True)
    
    # por ultimo, fiteamos las regresiones a los datos
    lm.fit(X_train, y_train)
    lm_ridge_cv.fit(X_train, y_train)
    lm_lasso_cv.fit(X_train, y_train)
    elastic_net.fit(X_train, y_train)
    
    return lm, lm_ridge_cv, lm_lasso_cv, elastic_net

Luego, definimos la función ```extraer_scores``` que, dado un modelo, devuelva su nombre, y el R2 y RMSE para train y test

def extraer_scores(modelo, X_train, X_test, y_train, y_test, variables):
    nombre = type(modelo)
    
    rmse = lambda y, y_pred: np.sqrt(mean_squared_error(y, y_pred))
    
    #datos para train
    R2_train = modelo.score(X_train, y_train)

    y_pred_tr_modelo = modelo.predict(X_train)
    RMSE_train = rmse(y_train,y_pred_tr_modelo)
    
    #datos para test
    R2_test = modelo.score(X_test, y_test)

    y_pred_modelo = modelo.predict(X_test)
    RMSE_test = rmse(y_test, y_pred_modelo)
    
    #el ouput será una Serie de pandas. Para facilitar la legibilidad, los juntamos primero en una lista
    datos = [nombre, R2_train, RMSE_train , R2_test, RMSE_test]
    return pd.Series(datos, index=variables)

#variables será una lista con los nombres de los datos que queremos saber
#en particular, está pensado para que sea ['Modelo', 'R2 train', 'RMSE train', 'R2 test', 'RMSE test']
#aparece en la prueba

La función ```df_scores``` devolverá el dataframe deseado, para el que cada fila representará un modelo particular con los datos que queremos saber

# el parámetro modelos será una lista de modelos
def df_scores(modelos, X_train, X_test, y_train, y_test, variables):
    dataframe = pd.DataFrame(columns=variables)
    
    for model in modelos:
        datos = extraer_scores(model, X_train, X_test, y_train, y_test, variables)
        dataframe = dataframe.append(datos, ignore_index=True)
        
    return dataframe

### Prueba 
Probamos la función ```df_scores``` para el DataFrame df_depto

#primero creamos los modelos llamando a la función regresiones
#pasamos como parámetros los conjuntos X_train e y_train que creamos más arriba
lm, lm_ridge_cv, lm_lasso_cv, elastic_net = regresiones(X_train, y_train)

#dado que el parámetro modelos de df_scores debe ser una lista, creamos una lista con los modelos
modelos = [lm, lm_ridge_cv, lm_lasso_cv, elastic_net]

#definimos también las variables que van a funcionar como columnas del dataframe
variables = ['Modelo', 'R2 train', 'RMSE train', 'R2 test', 'RMSE test']

#luego obtenemos el dataframe 
df_scores(modelos,X_train, X_test, y_train, y_test, variables)

# Regresiones por barrio

Ahora falta armar una función que tome los DataFrames en los barrios que tengan más de una cierta cantidad de registros. 
Antes de crear esta funcón, veremos cuántos registros hay por barrio, ciudad y provincia

La función ```print_barrios_a_calcular``` devuelve directamente la cantidad de registros por barrio, ciudad y provincia

def print_barrios_a_calcular(df):
    barrios = df.barrio.unique()
    ciudades = df.ciudad.unique()
    provincias = df.state_name.unique()
    
    print('Tamaño total del dataset: %d' % len(df))
    
    print('\n')
    
    print('Barrios con más de 750 registros')
    registros_barrio = 0 #acá se acumulará el total de registros por barrio
    for barrio in range(len(barrios)):
        mask_barrio = df.barrio == barrios[barrio]
        if len(df[mask_barrio]) > 750:
            print(barrios[barrio], len(df[mask_barrio]))
            registros_barrio += len(df[mask_barrio])
    print('Total: %d' % registros_barrio)
    print('Diferencia: %d' % (len(df) - registros_barrio))
    
    print('\n')
    
    print('Ciudades con más de 750 registros')
    registros_ciudad = 0 #total de registros por ciudad
    for ciudad in range(len(ciudades)):
        mask_ciudad = df.ciudad == ciudades[ciudad]
        if len(df[mask_ciudad]) > 750:
            print(ciudades[ciudad], len(df[mask_ciudad]))
            registros_ciudad += len(df[mask_ciudad])
    print('Total: %d' % registros_ciudad)
    print('Diferencia: %d' % (len(df) - registros_ciudad))
    
    print('\n')
    
    print('Provincias con más de 750 registros')
    registros_provincia = 0 #total de registros por provincia
    for provincia in range(len(provincias)):
        mask_provincia = df.state_name == provincias[provincia]
        if len(df[mask_provincia]) > 750:
            print(provincias[provincia], len(df[mask_provincia]))
            registros_provincia += len(df[mask_provincia])
    print('Total: %d' % registros_provincia)
    print('Diferencia: %d' % (len(df) - registros_provincia))

print_barrios_a_calcular(df_dpto)

Hubo poca diferencia entre los registros en provincias que tuviesen más de 750 registros. Veamos en qué provincias no se cumple la condición

provincias = df_dpto.state_name.unique()
indices_registros = []

for provincia in provincias:
    mask_provincia = df_dpto.state_name == provincia
    if len(df_dpto[mask_provincia]) < 750: #como se ve, la condición es la inversa a la que se pedía antes.
        print(provincia, len(df_dpto[mask_provincia]))

## Condicionales
Para los registros que no cumplen esa condición, se buscará la ciudad en la que están, y si la ciudad cumple la condición de la cantidad de registros. Por último, se hará el mismo proceso para provincia.<br>
En los casos en que la condición se cumpla, se usará el DataFrame resultante para entrenar un modelo. Para eso, se lo separará en train_test_split, y se llamará a la función ```regresiones ```

def train_por_zona(df, predictoras):
    barrios = df.barrio.unique()
    ciudades = df.ciudad.unique()
    provincias = df.state_name.unique()
    dataframe = pd.DataFrame()
    
    for barrio in barrios:
        mask_barrio = df.barrio == barrio
        
        #para cada barrio del df, evaluamos la condición de que tenga más de 750 registros
        if len(df[mask_barrio]) > 750:
            
            #si cumple la función, entonces separamos conjuntos de train y test sobre este df
            X = df.loc[mask_barrio, predictoras]
            y = df[mask_barrio].price_aprox_usd
            X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=123)
            
            #luego, armamos modelos con los conjuntos de train
            lm, lm_ridge_cv, lm_lasso_cv, elastic_net = regresiones(X_train, y_train)
            modelos = [lm, lm_ridge_cv, lm_lasso_cv, elastic_net]
            
            #creamos el índice complejo para el barrio
            iterables = [[barrio], ['Lineal', 'Ridge', 'Lasso', 'ElasticNet']]
            multi_indice = pd.MultiIndex.from_product(iterables, names=['Zona', 'Regresión'])
            
            #el df resultante tendrá una fila por modelo para este barrio
            dataframe_barrio = df_scores(modelos, X_train, X_test, y_train, y_test, variables)
            dataframe_barrio.index = multi_indice
            #luego, sumamos el df de este barrio al df común
            dataframe = dataframe.append(dataframe_barrio, ignore_index=False)
    
    
    #los mismos comentarios aplican al caso de ciudades y provincias
    
    for ciudad in ciudades:
        mask_ciudad = df.ciudad == ciudad
        
        if len(df[mask_ciudad]) > 750:
            X = df.loc[mask_ciudad, predictoras]
            y = df[mask_ciudad].price_aprox_usd
            X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=123)
            
            lm, lm_ridge_cv, lm_lasso_cv, elastic_net = regresiones(X_train, y_train)
            modelos = [lm, lm_ridge_cv, lm_lasso_cv, elastic_net]
            
            iterables = [[ciudad], ['Lineal', 'Ridge', 'Lasso', 'ElasticNet']]
            multi_indice = pd.MultiIndex.from_product(iterables, names=['Zona', 'Regresión'])
            
            dataframe_ciudad = df_scores(modelos, X_train, X_test, y_train, y_test, variables)
            dataframe_ciudad.index = multi_indice
            dataframe = dataframe.append(dataframe_ciudad, ignore_index=False)
    
    
    for provincia in provincias:
        mask_provincia = df.state_name == provincia
        
        if len(df[mask_provincia]) > 750:
            X = df.loc[mask_provincia, predictoras]
            y = df[mask_provincia].price_aprox_usd
            X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=123)
            
            lm, lm_ridge_cv, lm_lasso_cv, elastic_net = regresiones(X_train, y_train)
            modelos = [lm, lm_ridge_cv, lm_lasso_cv, elastic_net]
            
            iterables = [[provincia], ['Lineal', 'Ridge', 'Lasso', 'ElasticNet']]
            multi_indice = pd.MultiIndex.from_product(iterables, names=['Zona', 'Regresión'])
            
            dataframe_provincia = df_scores(modelos, X_train, X_test, y_train, y_test, variables)
            dataframe_provincia.index = multi_indice
            dataframe = dataframe.append(dataframe_provincia, ignore_index=False)
            
    dataframe.drop('Modelo', axis=1)
    return dataframe

df_final = train_por_zona(df_dpto, predictoras)

df_final.drop('Modelo', axis=1)

## Algunas observaciones sobre df_final 

### Máximos y mínimos 

En cuanto a los maximos valores estuvieron parejas

print('Train')
print(df_final['R2 train'].max(level='Regresión'))
print('\n')
print('Test')
print(df_final['R2 test'].max(level='Regresión'))

Hubo más diferencia en las peores. Pero de todos modos, los resultados no son muy malos

print('Train')
print(df_final['R2 train'].min(level='Regresión'))
print('\n')
print('Test')
print(df_final['R2 test'].min(level='Regresión'))

print(max(df_final['R2 train']))
print(max(df_final['R2 test']))
print(min(df_final['R2 train']))
print(min(df_final['R2 test']))

### Media de los scores

Promedio de los scores para train y test

print(df_final['R2 train'].values.mean())
print(df_final['R2 test'].values.mean())

### DFs por barrio, ciudad y provincia 

barrios = [x for x in df_dpto.barrio.unique() if len(df_dpto[df_dpto.barrio == x]) > 750]
ciudad = [x for x in df_dpto.ciudad.unique() if len(df_dpto[df_dpto.ciudad == x]) > 750]
provincias = [x for x in df_dpto.state_name.unique() if len(df_dpto[df_dpto.state_name == x]) > 750]

df_final.loc[barrios]

df_final.loc[ciudad]

df_final.loc[provincias]

set(df_final.loc[ciudad].index.levels[0]) - set(df_final.loc[barrios].index.levels[0])

Provincia, ciudad y barrios tienen la misma cantidad de índices. Es decir, hay repetidos. Pero hay diferencias

print(len(set(df_final.loc[ciudad].index.levels[0])))
print(len(set(df_final.loc[barrios].index.levels[0])))
print(len(set(df_final.loc[provincias].index.levels[0])))
